requirements.txt
pytest == 7.2.0
selenium == 4.6.0
mysql-connector-python == 8.0.31
request == 2.28.1